+++ CHANGELOG.rst
@@ -2,12 +2,20 @@
 Changelog
 =========
 
+Version 10.11
+=============
+[Release Date]
+
+- Similar functionality to version 10.10 but the GUI code has been merged into the same repository
+  as the command-line/library code.
+
 Version 10.10
 =============
 25 October 2023
- - Sensitivty sweep: change limits on GUI inputs so that this is suitable for background sweep,
-   add noise handling.
- - Bugfix: disable fitting column width for Datetime column, as this sometimes causes crashes.
+
+- Sensitivty sweep: change limits on GUI inputs so that this is suitable for background sweep,
+  add noise handling.
+- Bugfix: disable fitting column width for Datetime column, as this sometimes causes crashes.
 
 
 Version 10.9